+++ src/redux/notices/noticeOperations.js
@@ -97,10 +97,10 @@ export const deleteNoticeThunk = createAsyncThunk(
 
 export const getSelectedNoticeThunk = createAsyncThunk(
   'notices/noticeById',
-  async ({noticeId, thunk}, {rejectedWithValue, dispatch} ) => {
+  async ({id}, {rejectedWithValue, dispatch} ) => {
     try {
-      const { data } = await noticeInstance.get(`/api/notices/${noticeId}`);
-      dispatch(thunk);
+      const { data } = await authInstance.get(`/api/notices/${id}`);
+     
       return data; 
     } catch (error) {
       return rejectedWithValue(error.message);
